+++ docs/conf.py
@@ -26,14 +26,12 @@ release = "0.0.5"
 
 master_doc = "index"
 
-# Add any Sphinx extension module names here, as strings. They can be
-# extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
-# ones.
+# Add any Sphinx extension module names here, as strings. They can be extensions coming
+# with Sphinx (named 'sphinx.ext.*') or your custom ones.
 extensions = [
     "numpydoc",
     "sphinx.ext.extlinks",
     "sphinx.ext.intersphinx",
-    "sphinx_rtd_theme",
 ]
 
 # Add any paths that contain templates here, relative to this directory.
@@ -48,6 +46,9 @@ exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
 
 # -- Extensions configuration ----------------------------------------------------------
 
+# Configuration for autodoc
+autosummary_generate = True
+
 extlinks = {
     "ghuser": ("https://github.com/%s", "@"),
     "gh": ("https://github.com/OpenSourceEconomics/pipeline/pulls/%s", "#"),
@@ -59,16 +60,25 @@ intersphinx_mapping = {
     "python": ("https://docs.python.org/3.7", None),
 }
 
+# Configuration for numpydoc
+numpydoc_xref_param_type = True
+numpydoc_xref_ignore = {"type", "optional", "default"}
+
 
 # -- Options for HTML output -----------------------------------------------------------
 
 # The theme to use for HTML and HTML Help pages.  See the documentation for a list of
-# builtin themes.
+# built-in themes.
 
-html_theme = "sphinx_rtd_theme"
+html_theme = "pydata_sphinx_theme"
 
 # Add any paths that contain custom static files (such as style sheets) here, relative
-# to this directory. They are copied after the builtin static files, so a file named
-# "default.css" will overwrite the builtin "default.css".
+# to this directory. They are copied after the built-in static files, so a file named
+# "default.css" will overwrite the built-in "default.css".
+html_css_files = ["css/custom.css"]
 
-# html_static_path = ["_static"]  # noqa: E800
+html_static_path = ["_static"]
+
+html_theme_options = {
+    "github_url": "https://github.com/OpenSourceEconomics/pipeline",
+}