+++ pipeline/database.py
@@ -12,14 +12,6 @@ class Hash(db.Entity):
     orm.PrimaryKey(task, dependency)
 
 
-class TaskOutput(db.Entity):
-    task = orm.Required(str)
-    target = orm.Required(str)
-    data = orm.Required(orm.buffer)
-
-    orm.PrimaryKey(task, target)
-
-
 def create_database(config):
     db.bind(**config["db"])
     db.generate_mapping(create_tables=True)