+++ pipeline/hashing.py
@@ -41,12 +41,9 @@ def compare_hashes_of_task(id_, env, dag, config):
     for node in dependencies_and_targets:
         path = Path(node)
 
-        if node in env.list_templates() or path.exists():
-            if node in env.list_templates():
-                rendered_task = render_task_template(id_, dag.nodes[id_], env, config)
-                hash_ = _compute_hash_of_string(rendered_task)
-            else:
-                hash_ = _compute_hash_of_file(path, path.stat().st_mtime)
+        if node in env.list_templates():
+            rendered_task = render_task_template(id_, dag.nodes[id_], env, config)
+            hash_ = _compute_hash_of_string(rendered_task)
 
             try:
                 hash_in_db = Hash[id_, node]
@@ -57,6 +54,22 @@ def compare_hashes_of_task(id_, env, dag, config):
             except AssertionError:
                 hash_in_db.hash_ = hash_
                 have_same_hashes = False
+
+        elif path.exists():
+            paths = _path_to_file_or_directory_to_path_iterator(path)
+
+            for path in paths:
+                hash_ = _compute_hash_of_file(path, path.stat().st_mtime)
+
+                try:
+                    hash_in_db = Hash[path.as_posix(), node]
+                    assert hash_ == hash_in_db.hash_
+                except orm.ObjectNotFound:
+                    Hash(task=id_, dependency=node, hash_=hash_)
+                    have_same_hashes = False
+                except AssertionError:
+                    hash_in_db.hash_ = hash_
+                    have_same_hashes = False
 
         else:
             have_same_hashes = False
@@ -71,31 +84,43 @@ def save_hashes_of_task_dependencies(id_, env, dag, config):
         if dependency in env.list_templates():
             rendered_task = render_task_template(id_, dag.nodes[id_], env, config)
             hash_ = _compute_hash_of_string(rendered_task)
-        else:
-            path = Path(dependency)
-            hash_ = _compute_hash_of_file(path, path.stat().st_mtime)
 
-        try:
-            hash_in_db = Hash[id_, dependency]
-        except orm.ObjectNotFound:
-            Hash(task=id_, dependency=dependency, hash_=hash_)
+            try:
+                hash_in_db = Hash[id_, dependency]
+            except orm.ObjectNotFound:
+                Hash(task=id_, dependency=dependency, hash_=hash_)
+            else:
+                hash_in_db.hash = hash_
+
         else:
-            hash_in_db.hash = hash_
+            paths = _path_to_file_or_directory_to_path_iterator(dependency)
+
+            for p in paths:
+                hash_ = _compute_hash_of_file(p, p.stat().st_mtime)
+
+                try:
+                    hash_in_db = Hash[id_, p.as_posix()]
+                except orm.ObjectNotFound:
+                    Hash(task=id_, dependency=dependency, hash_=hash_)
+                else:
+                    hash_in_db.hash = hash_
 
 
 @orm.db_session
 def save_hash_of_task_target(id_, dag):
     """Loop over the targets of a task and save the hashes of the files."""
-    for string_path in ensure_list(dag.nodes[id_]["produces"]):
-        path = Path(string_path)
-        hash_ = _compute_hash_of_file(path, path.stat().st_mtime)
+    for path in ensure_list(dag.nodes[id_]["produces"]):
+        paths = _path_to_file_or_directory_to_path_iterator(path)
 
-        try:
-            hash_in_db = Hash[id_, string_path]
-        except orm.ObjectNotFound:
-            Hash(task=id_, dependency=string_path, hash_=hash_)
-        else:
-            hash_in_db.hash = hash_
+        for p in paths:
+            hash_ = _compute_hash_of_file(p, p.stat().st_mtime)
+
+            try:
+                hash_in_db = Hash[id_, p]
+            except orm.ObjectNotFound:
+                Hash(task=id_, dependency=p.as_posix(), hash_=hash_)
+            else:
+                hash_in_db.hash = hash_
 
 
 @functools.lru_cache()  # noqa: U101
@@ -124,8 +149,26 @@ def _compute_hash_of_file(path, _last_modified=None, algorithm="sha256"):
 
 
 def _compute_hash_of_string(string, algorithm="sha256"):
-    """Compute hash of a string."""
+    """Compute hash of a string.
+
+    Example
+    -------
+    >>> _compute_hash_of_string("abcde")
+    '36bbe50ed96841d10443bcb670d6554f0a34b761be67ec9c4a8ad2c0c44ca42c'
+
+    """
     h = hashlib.new(algorithm)
     h.update(string.encode("utf-8"))
 
     return h.hexdigest()
+
+
+def _path_to_file_or_directory_to_path_iterator(path):
+    """Convert a path to a file or directory to an iterator over paths."""
+    path = Path(path)
+    if path.is_dir():
+        paths = list(path.rglob("*"))
+    else:
+        paths = [path]
+
+    return paths