+++ docs/release_notes.rst
@@ -13,7 +13,10 @@ releases are available on `Anaconda.org
 - Add ``run_always`` to the task dictionary to always execute a task (:gh:`12`).
 - Add project root directory to the ``PYTHONPATH`` so that every task can import with
   ``from src.... import ...`` (:gh:`13`).
+- Split coverage reports into unit, integration, and end-to-end tests (:gh:`14`).
 - Allow to have directories as task dependencies and targets (:gh:`15`).
+- Handle multiple targets of tasks (:gh:`18`).
+- Fix exception handling in the parallel executor (:gh:`19`).
 
 
 0.0.5 - 2020-04-26